move tools used by the notebook into barrier3d.tools
